+++ apps/fountain/lib/fountain_web/controllers/account_data_controller.ex
@@ -24,7 +24,9 @@ defmodule FountainWeb.AccountDataController do
 
   action_fallback FountainWeb.FallbackController
 
-  plug OpenApiSpex.Plug.CastAndValidate, replace_params: false
+  plug OpenApiSpex.Plug.CastAndValidate,
+    replace_params: false,
+    render_error: FountainWeb.Plugs.CastRenderError
 
   tags(["Account"])