+++ apps/fountain/lib/fountain_web/controllers/audit_controller.ex
@@ -19,7 +19,9 @@ defmodule FountainWeb.AuditController do
 
   action_fallback FountainWeb.FallbackController
 
-  plug OpenApiSpex.Plug.CastAndValidate, replace_params: false
+  plug OpenApiSpex.Plug.CastAndValidate,
+    replace_params: false,
+    render_error: FountainWeb.Plugs.CastRenderError
 
   @default_limit 100
   @max_limit 500