+++ apps/fountain/lib/fountain_web/controllers/buzz_agent_controller.ex
@@ -20,7 +20,9 @@ defmodule FountainWeb.BuzzAgentController do
 
   action_fallback FountainWeb.FallbackController
 
-  plug OpenApiSpex.Plug.CastAndValidate, replace_params: false
+  plug OpenApiSpex.Plug.CastAndValidate,
+    replace_params: false,
+    render_error: FountainWeb.Plugs.CastRenderError
 
   tags(["Buzz"])