+++ apps/fountain/lib/fountain_web/controllers/environment_controller.ex
@@ -10,7 +10,9 @@ defmodule FountainWeb.EnvironmentController do
 
   action_fallback FountainWeb.FallbackController
 
-  plug OpenApiSpex.Plug.CastAndValidate, replace_params: false
+  plug OpenApiSpex.Plug.CastAndValidate,
+    replace_params: false,
+    render_error: FountainWeb.Plugs.CastRenderError
 
   tags(["Environments"])