+++ apps/fountain/lib/fountain_web/controllers/sandbox_controller.ex
@@ -17,7 +17,9 @@ defmodule FountainWeb.SandboxController do
 
   action_fallback FountainWeb.FallbackController
 
-  plug OpenApiSpex.Plug.CastAndValidate, replace_params: false
+  plug OpenApiSpex.Plug.CastAndValidate,
+    replace_params: false,
+    render_error: FountainWeb.Plugs.CastRenderError
 
   tags(["Sandboxes"])