+++ apps/fountain/lib/fountain_web/controllers/search_controller.ex
@@ -13,7 +13,9 @@ defmodule FountainWeb.SearchController do
 
   action_fallback FountainWeb.FallbackController
 
-  plug OpenApiSpex.Plug.CastAndValidate, replace_params: false
+  plug OpenApiSpex.Plug.CastAndValidate,
+    replace_params: false,
+    render_error: FountainWeb.Plugs.CastRenderError
 
   tags(["Search"])